+++ btgsolutions_dataservices/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "4.4.0"
+__version__ = "4.4.1"
 
 from .websocket import *
 from .rest import *